+++ examples/4337-gas-metering/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@safe-global/4337-gas-metering",
+  "name": "@safe-global/safe-modules-example-4337-gas-metering",
   "version": "1.0.0",
   "description": "Gas Metering Analysis for Safe with 4337 Module",
   "homepage": "https://github.com/safe-global/safe-modules/4337-gas-metering",
