+++ examples/4337-passkeys/README.md
@@ -33,7 +33,7 @@ Helpful links:
 ### Run the app in development mode
 
 ```bash
-npm run dev -w examples/safe-4337-passkeys
+npm run dev -w examples/4337-passkeys
 ```
 
 ## Config adjustments
