+++ examples/4337-passkeys/src/components/SafeCard.tsx
