+++ examples/4337-passkeys/src/utils.ts
