+++ modules/4337/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@safe-global/safe-erc4337",
+  "name": "@safe-global/safe-4337",
   "version": "0.3.0",
   "description": "Safe Module for ERC-4337 support",
   "homepage": "https://github.com/safe-global/safe-modules/tree/main/modules/4337",
