+++ modules/passkey/package.json
@@ -43,7 +43,7 @@
     "@account-abstraction/contracts": "^0.7.0",
     "@noble/curves": "^1.3.0",
     "@nomicfoundation/hardhat-toolbox": "^4.0.0",
-    "@safe-global/safe-erc4337": "^0.3.0",
+    "@safe-global/safe-4337": "^0.3.0",
     "@simplewebauthn/server": "^9.0.3",
     "cbor": "^9.0.2",
     "dotenv": "^16.4.5",
